refactor: move files in `utils` into `common`
