+++ t4-devkit/t4_devkit/tier4.py
@@ -10,8 +10,8 @@ from nuscenes.nuscenes import LidarPointCloud, RadarPointCloud
 from pyquaternion import Quaternion
 import rerun as rr
 import rerun.blueprint as rrb
+from t4_devkit.common import Box2D, Box3D, is_box_in_image
 from t4_devkit.schema import SchemaName, SensorModality, VisibilityLevel, build_schema
-from t4_devkit.utils import Box2D, Box3D, is_box_in_image
 
 if TYPE_CHECKING:
     from t4_devkit.typing import CamIntrinsicType, VelocityType
@@ -603,6 +603,7 @@ class Tier4:
         first_radar_tokens: list[str] = []
         first_camera_tokens: list[str] = []
 
+        # FIXME: if sample data is not associated with the first sample, all frames are not rendered.
         for channel, sd_token in first_sample.data.items():
             self._render_sensor_calibration(sd_token)
             if channel.modality == SensorModality.LIDAR:
@@ -698,6 +699,7 @@ class Tier4:
             first_camera_tokens (list[str]): List of first sample data tokens corresponding to cameras.
             max_timestamp_us (float): Max time length in [us].
         """
+        print(len(first_camera_tokens))
         for first_camera_token in first_camera_tokens:
             current_camera_token = first_camera_token
             while current_camera_token != "":
