--- t4-devkit/t4_devkit/utils/__init__.py
@@ -1,2 +1,0 @@
-from .data_classes import *  # noqa
-from .geometry import *  # noqa
